refactor: Calculator컴포넌트 container로 이동
