+++ calculator/src/App.js
@@ -1,5 +1,5 @@
 import { Component } from 'react';
-import Calculator from './component/Calculator';
+import Calculator from './container/Calculator';
 import './App.css';
 export default class App extends Component {
   render() {
